fix to restore symmetry between storing and restoring the value of #defines. Restoring was truncated to an int on a 64 bit machine.
git-svn-id: http://root.cern.ch/svn/root/trunk@7216 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/src/var.c
@@ -4508,7 +4508,7 @@ struct G__var_array *varglobal,*varlocal;
 	}
 	else {
 	  G__letint(&result,'i',
-		    *(int *)(G__struct_offset+var->p[ig15]+p_inc*G__LONGALLOC));
+		    *(long *)(G__struct_offset+var->p[ig15]+p_inc*G__LONGALLOC));
 	}
 	break;
       }
